Organize unbounded tests into namespaces
