+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/IbmDb2/IbmDb2AsyncTests.cs
@@ -11,13 +11,13 @@
 //using Xunit;
 //using Xunit.Abstractions;
 
-//namespace NewRelic.Agent.UnboundedIntegrationTests
+//namespace NewRelic.Agent.UnboundedIntegrationTests.IbmDb2
 //{
-//    public class IbmDb2Tests : IClassFixture<RemoteServiceFixtures.IbmDb2BasicMvcFixture>
+//    public class IbmDb2AsyncTests : IClassFixture<RemoteServiceFixtures.IbmDb2BasicMvcFixture>
 //    {
 //        private readonly RemoteServiceFixtures.IbmDb2BasicMvcFixture _fixture;
 
-//        public IbmDb2Tests(RemoteServiceFixtures.IbmDb2BasicMvcFixture fixture, ITestOutputHelper output)
+//        public IbmDb2AsyncTests(RemoteServiceFixtures.IbmDb2BasicMvcFixture fixture, ITestOutputHelper output)
 //        {
 //            _fixture = fixture;
 //            _fixture.TestLogger = output;
@@ -37,7 +37,7 @@
 //                },
 //                exerciseApplication: () =>
 //                {
-//                    _fixture.GetIbmDb2();
+//                    _fixture.GetIbmDb2Async();
 //                }
 //            );
 //            _fixture.Initialize();
@@ -53,24 +53,24 @@
 //                new Assertions.ExpectedMetric { metricName = @"Datastore/allWeb", callCount = 4 },
 //                new Assertions.ExpectedMetric { metricName = @"Datastore/IBMDB2/all", callCount = 4 },
 //                new Assertions.ExpectedMetric { metricName = @"Datastore/IBMDB2/allWeb", callCount = 4 },
-//                new Assertions.ExpectedMetric { metricName = $"Datastore/instance/IBMDB2/{CommonUtils.NormalizeHostname(Db2Configuration.Db2Server)}/default", callCount = 4},
+//                new Assertions.ExpectedMetric { metricName = $"Datastore/instance/IBMDB2/{CommonUtils.NormalizeHostname(CommonUtils.NormalizeHostname(Db2Configuration.Db2Server))}/default", callCount = 4},
 //                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/IBMDB2/select", callCount = 2 },
 //                new Assertions.ExpectedMetric { metricName = @"Datastore/statement/IBMDB2/employee/select", callCount = 1 },
-//                new Assertions.ExpectedMetric { metricName = @"Datastore/statement/IBMDB2/employee/select", callCount = 1, metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2Query"},
+//                new Assertions.ExpectedMetric { metricName = @"Datastore/statement/IBMDB2/employee/select", callCount = 1, metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2QueryAsync"},
 //                //ExecuteScalar() double instrumented: DOTNET-1800
 //                new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/IBMDB2/{_fixture.TableName}/select", callCount = 1 },
-//                new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/IBMDB2/{_fixture.TableName}/select", callCount = 1, metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2Query"},
+//                new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/IBMDB2/{_fixture.TableName}/select", callCount = 1, metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2QueryAsync"},
 //                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/IBMDB2/insert", callCount = 1 },
 //                new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/IBMDB2/{_fixture.TableName}/insert", callCount = 1 },
-//                new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/IBMDB2/{_fixture.TableName}/insert", callCount = 1, metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2Query"},
+//                new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/IBMDB2/{_fixture.TableName}/insert", callCount = 1, metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2QueryAsync"},
 //                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/IBMDB2/delete", callCount = 1 },
 //                new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/IBMDB2/{_fixture.TableName}/delete", callCount = 1 },
-//                new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/IBMDB2/{_fixture.TableName}/delete", callCount = 1, metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2Query"},
+//                new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/IBMDB2/{_fixture.TableName}/delete", callCount = 1, metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2QueryAsync"},
 
 //                // We are not checking callCount on Iterate metrics because they can be confusing in that calls like Open can result in calls to Read.
 //                // This is particularly true for MySQL, but doing this for all vendors for consistency.
 //                new Assertions.ExpectedMetric { metricName = @"DotNet/DatabaseResult/Iterate" },
-//                new Assertions.ExpectedMetric { metricName = @"DotNet/DatabaseResult/Iterate", metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2Query"}
+//                new Assertions.ExpectedMetric { metricName = @"DotNet/DatabaseResult/Iterate", metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2QueryAsync"}
 //            };
 //            var unexpectedMetrics = new List<Assertions.ExpectedMetric>
 //            {
@@ -79,9 +79,9 @@
 //                new Assertions.ExpectedMetric { metricName = @"Datastore/IBMDB2/allOther"},
 
 //                // The operation metric should not be scoped because the statement metric is scoped instead
-//                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/IBMDB2/select", metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2Query" },
-//                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/IBMDB2/insert", metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2Query" },
-//                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/IBMDB2/delete", metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2Query" }
+//                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/IBMDB2/select", metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2QueryAsync" },
+//                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/IBMDB2/insert", metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2QueryAsync" },
+//                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/IBMDB2/delete", metricScope = "WebTransaction/MVC/DefaultController/InvokeIbmDb2QueryAsync" }
 //            };
 //            var expectedTransactionTraceSegments = new List<string>
 //            {
@@ -104,14 +104,14 @@
 //            {
 //                new Assertions.ExpectedSqlTrace
 //                {
-//                    TransactionName = "WebTransaction/MVC/DefaultController/InvokeIbmDb2Query",
+//                    TransactionName = "WebTransaction/MVC/DefaultController/InvokeIbmDb2QueryAsync",
 //                    Sql = "SELECT LASTNAME FROM EMPLOYEE FETCH FIRST ROW ONLY",
 //                    DatastoreMetricName = "Datastore/statement/IBMDB2/employee/select",
 //                    HasExplainPlan = false
 //                },
 //                new Assertions.ExpectedSqlTrace
 //                {
-//                    TransactionName = "WebTransaction/MVC/DefaultController/InvokeIbmDb2Query",
+//                    TransactionName = "WebTransaction/MVC/DefaultController/InvokeIbmDb2QueryAsync",
 //                    Sql = $"SELECT COUNT(*) FROM {_fixture.TableName}",
 //                    DatastoreMetricName = $"Datastore/statement/IBMDB2/{_fixture.TableName}/select",
 
@@ -119,7 +119,7 @@
 //                },
 //                new Assertions.ExpectedSqlTrace
 //                {
-//                    TransactionName = "WebTransaction/MVC/DefaultController/InvokeIbmDb2Query",
+//                    TransactionName = "WebTransaction/MVC/DefaultController/InvokeIbmDb2QueryAsync",
 //                    Sql = $"INSERT INTO {_fixture.TableName} (HOTEL_ID, BOOKING_DATE) VALUES (?, SYSDATE)",
 //                    DatastoreMetricName = $"Datastore/statement/IBMDB2/{_fixture.TableName}/insert",
 
@@ -127,7 +127,7 @@
 //                },
 //                new Assertions.ExpectedSqlTrace
 //                {
-//                    TransactionName = "WebTransaction/MVC/DefaultController/InvokeIbmDb2Query",
+//                    TransactionName = "WebTransaction/MVC/DefaultController/InvokeIbmDb2QueryAsync",
 //                    Sql = $"DELETE FROM {_fixture.TableName} WHERE HOTEL_ID = ?",
 //                    DatastoreMetricName = $"Datastore/statement/IBMDB2/{_fixture.TableName}/delete",
 
@@ -136,8 +136,8 @@
 //            };
 
 //            var metrics = _fixture.AgentLog.GetMetrics().ToList();
-//            var transactionSample = _fixture.AgentLog.TryGetTransactionSample("WebTransaction/MVC/DefaultController/InvokeIbmDb2Query");
-//            var transactionEvent = _fixture.AgentLog.TryGetTransactionEvent("WebTransaction/MVC/DefaultController/InvokeIbmDb2Query");
+//            var transactionSample = _fixture.AgentLog.TryGetTransactionSample("WebTransaction/MVC/DefaultController/InvokeIbmDb2QueryAsync");
+//            var transactionEvent = _fixture.AgentLog.TryGetTransactionEvent("WebTransaction/MVC/DefaultController/InvokeIbmDb2QueryAsync");
 //            var sqlTraces = _fixture.AgentLog.GetSqlTraces().ToList();
 
 //            NrAssert.Multiple(
