+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/MsSql/MsSqlAsyncLegacyAspPipelineTests.cs
@@ -10,7 +10,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.MsSql
 {
     public class MsSqlAsyncLegacyAspPipelineTests : IClassFixture<RemoteServiceFixtures.MsSqlBasicMvcFixture>
     {
