+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/Msmq/MsmqConsumeTests.cs
@@ -8,13 +8,20 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.Msmq
 {
-    public class MsmqPurgeTests : IClassFixture<RemoteServiceFixtures.MSMQBasicMVCApplicationFixture>
+    /// <summary>
+    /// Integration Test for MSMQ Consume message.
+    /// </summary>
+    /// <remarks>
+    /// Although this test also uses the MSMQ Send endpoint, it is important to keep the Send and Consume tests as separate fixtures
+    /// in order to separately test the existence of a Consume segment in the transaction trace. Only a single TT is being saved per test.
+    /// </remarks>
+    public class MsmqConsumeTests : IClassFixture<RemoteServiceFixtures.MSMQBasicMVCApplicationFixture>
     {
         private readonly RemoteServiceFixtures.MSMQBasicMVCApplicationFixture _fixture;
 
-        public MsmqPurgeTests(RemoteServiceFixtures.MSMQBasicMVCApplicationFixture fixture, ITestOutputHelper output)
+        public MsmqConsumeTests(RemoteServiceFixtures.MSMQBasicMVCApplicationFixture fixture, ITestOutputHelper output)
         {
             _fixture = fixture;
             _fixture.TestLogger = output;
@@ -28,7 +35,7 @@ namespace NewRelic.Agent.UnboundedIntegrationTests
                 exerciseApplication: () =>
                 {
                     _fixture.GetMessageQueue_Msmq_Send(true);
-                    _fixture.GetMessageQueue_Msmq_Purge();
+                    _fixture.GetMessageQueue_Msmq_Receive();
                 }
             );
             _fixture.Initialize();
@@ -39,17 +46,17 @@ namespace NewRelic.Agent.UnboundedIntegrationTests
         {
             var expectedMetrics = new List<Assertions.ExpectedMetric>
             {
-                new Assertions.ExpectedMetric { metricName = @"MessageBroker/Msmq/Queue/Purge/Named/private$\nrtestqueue", callCount = 1},
-                new Assertions.ExpectedMetric { metricName = @"MessageBroker/Msmq/Queue/Purge/Named/private$\nrtestqueue", callCount = 1, metricScope = "WebTransaction/MVC/MSMQController/Msmq_Purge"},
+                new Assertions.ExpectedMetric { metricName = @"MessageBroker/Msmq/Queue/Consume/Named/private$\nrtestqueue", callCount = 1},
+                new Assertions.ExpectedMetric { metricName = @"MessageBroker/Msmq/Queue/Consume/Named/private$\nrtestqueue", callCount = 1, metricScope = "WebTransaction/MVC/MSMQController/Msmq_Receive"}
             };
             var expectedTransactionTraceSegments = new List<string>
             {
-                @"MessageBroker/Msmq/Queue/Purge/Named/private$\nrtestqueue"
+                @"MessageBroker/Msmq/Queue/Consume/Named/private$\nrtestqueue"
             };
 
             var metrics = _fixture.AgentLog.GetMetrics().ToList();
-            var transactionSample = _fixture.AgentLog.TryGetTransactionSample("WebTransaction/MVC/MSMQController/Msmq_Purge");
-            var transactionEvent = _fixture.AgentLog.TryGetTransactionEvent("WebTransaction/MVC/MSMQController/Msmq_Purge");
+            var transactionSample = _fixture.AgentLog.TryGetTransactionSample("WebTransaction/MVC/MSMQController/Msmq_Receive");
+            var transactionEvent = _fixture.AgentLog.TryGetTransactionEvent("WebTransaction/MVC/MSMQController/Msmq_Receive");
 
             NrAssert.Multiple(
                 () => Assert.NotNull(transactionSample),
@@ -62,6 +69,5 @@ namespace NewRelic.Agent.UnboundedIntegrationTests
                 () => Assertions.TransactionTraceSegmentsExist(expectedTransactionTraceSegments, transactionSample)
             );
         }
-
     }
 }
