+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/Msmq/MsmqPeekTests.cs
@@ -8,7 +8,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.Msmq
 {
     public class MsmqPeekTests : IClassFixture<RemoteServiceFixtures.MSMQBasicMVCApplicationFixture>
     {
