+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/Msmq/MsmqPurgeTests.cs
@@ -8,13 +8,13 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.Msmq
 {
-    public class MsmqSendTests : IClassFixture<RemoteServiceFixtures.MSMQBasicMVCApplicationFixture>
+    public class MsmqPurgeTests : IClassFixture<RemoteServiceFixtures.MSMQBasicMVCApplicationFixture>
     {
         private readonly RemoteServiceFixtures.MSMQBasicMVCApplicationFixture _fixture;
 
-        public MsmqSendTests(RemoteServiceFixtures.MSMQBasicMVCApplicationFixture fixture, ITestOutputHelper output)
+        public MsmqPurgeTests(RemoteServiceFixtures.MSMQBasicMVCApplicationFixture fixture, ITestOutputHelper output)
         {
             _fixture = fixture;
             _fixture.TestLogger = output;
@@ -27,7 +27,8 @@ namespace NewRelic.Agent.UnboundedIntegrationTests
                 },
                 exerciseApplication: () =>
                 {
-                    _fixture.GetMessageQueue_Msmq_Send(false);
+                    _fixture.GetMessageQueue_Msmq_Send(true);
+                    _fixture.GetMessageQueue_Msmq_Purge();
                 }
             );
             _fixture.Initialize();
@@ -38,17 +39,17 @@ namespace NewRelic.Agent.UnboundedIntegrationTests
         {
             var expectedMetrics = new List<Assertions.ExpectedMetric>
             {
-                new Assertions.ExpectedMetric { metricName = @"MessageBroker/Msmq/Queue/Produce/Named/private$\nrtestqueue", callCount = 1},
-                new Assertions.ExpectedMetric { metricName = @"MessageBroker/Msmq/Queue/Produce/Named/private$\nrtestqueue", callCount = 1, metricScope = "WebTransaction/MVC/MSMQController/Msmq_Send"},
+                new Assertions.ExpectedMetric { metricName = @"MessageBroker/Msmq/Queue/Purge/Named/private$\nrtestqueue", callCount = 1},
+                new Assertions.ExpectedMetric { metricName = @"MessageBroker/Msmq/Queue/Purge/Named/private$\nrtestqueue", callCount = 1, metricScope = "WebTransaction/MVC/MSMQController/Msmq_Purge"},
             };
             var expectedTransactionTraceSegments = new List<string>
             {
-                @"MessageBroker/Msmq/Queue/Produce/Named/private$\nrtestqueue"
+                @"MessageBroker/Msmq/Queue/Purge/Named/private$\nrtestqueue"
             };
 
             var metrics = _fixture.AgentLog.GetMetrics().ToList();
-            var transactionSample = _fixture.AgentLog.TryGetTransactionSample("WebTransaction/MVC/MSMQController/Msmq_Send");
-            var transactionEvent = _fixture.AgentLog.TryGetTransactionEvent("WebTransaction/MVC/MSMQController/Msmq_Send");
+            var transactionSample = _fixture.AgentLog.TryGetTransactionSample("WebTransaction/MVC/MSMQController/Msmq_Purge");
+            var transactionEvent = _fixture.AgentLog.TryGetTransactionEvent("WebTransaction/MVC/MSMQController/Msmq_Purge");
 
             NrAssert.Multiple(
                 () => Assert.NotNull(transactionSample),
@@ -61,5 +62,6 @@ namespace NewRelic.Agent.UnboundedIntegrationTests
                 () => Assertions.TransactionTraceSegmentsExist(expectedTransactionTraceSegments, transactionSample)
             );
         }
+
     }
 }
