+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/NServiceBus/NServiceBusTests.cs
@@ -8,7 +8,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.NServiceBus
 {
     public class NServiceBusTests : IClassFixture<RemoteServiceFixtures.NServiceBusBasicMvcApplicationFixture>
     {
