+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/Oracle/OracleAsyncTests.cs
@@ -10,13 +10,13 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.Oracle
 {
-    public class EnterpriseLibraryOracleTests : IClassFixture<RemoteServiceFixtures.OracleBasicMvcFixture>
+    public class OracleAsyncTests : IClassFixture<RemoteServiceFixtures.OracleBasicMvcFixture>
     {
         private readonly RemoteServiceFixtures.OracleBasicMvcFixture _fixture;
 
-        public EnterpriseLibraryOracleTests(RemoteServiceFixtures.OracleBasicMvcFixture fixture, ITestOutputHelper output)
+        public OracleAsyncTests(RemoteServiceFixtures.OracleBasicMvcFixture fixture, ITestOutputHelper output)
         {
             _fixture = fixture;
             _fixture.TestLogger = output;
@@ -31,12 +31,14 @@ namespace NewRelic.Agent.UnboundedIntegrationTests
 
                     CommonUtils.ModifyOrCreateXmlAttributeInNewRelicConfig(configPath, new[] { "configuration", "transactionTracer" }, "explainThreshold", "1");
 
-                    var instrumentationFilePath = $@"{fixture.DestinationNewRelicExtensionsDirectoryPath}\NewRelic.Providers.Wrapper.Sql.Instrumentation.xml";
-                    CommonUtils.SetAttributeOnTracerFactoryInNewRelicInstrumentation(instrumentationFilePath, "", "enabled", "true");
+                    var instrumentationFilePath = $@"{fixture.DestinationNewRelicExtensionsDirectoryPath}\NewRelic.Providers.Wrapper.SqlAsync.Instrumentation.xml";
+                    CommonUtils.SetAttributeOnTracerFactoryInNewRelicInstrumentation(
+                       instrumentationFilePath,
+                        "", "enabled", "true");
                 },
                 exerciseApplication: () =>
                 {
-                    _fixture.GetEnterpriseLibraryOracle();
+                    _fixture.GetOracleAsync();
                 }
             );
             _fixture.Initialize();
@@ -54,36 +56,34 @@ namespace NewRelic.Agent.UnboundedIntegrationTests
                 new Assertions.ExpectedMetric { metricName = @"Datastore/Oracle/allWeb", callCount = 6 },
                 new Assertions.ExpectedMetric { metricName = $@"Datastore/instance/Oracle/{OracleConfiguration.OracleServer}/{OracleConfiguration.OraclePort}", callCount = 6},
                 new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Oracle/declare", callCount = 1 },
-                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Oracle/select", callCount = 3},
+                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Oracle/select", callCount = 3 },
                 new Assertions.ExpectedMetric { metricName = @"Datastore/statement/Oracle/user_tables/select", callCount = 1 },
-                new Assertions.ExpectedMetric { metricName = @"Datastore/statement/Oracle/user_tables/select", callCount = 1, metricScope = "WebTransaction/MVC/DefaultController/EnterpriseLibraryOracle"},
+                new Assertions.ExpectedMetric { metricName = @"Datastore/statement/Oracle/user_tables/select", callCount = 1, metricScope = "WebTransaction/MVC/DefaultController/OracleAsync"},
                 //ExecuteScalar() double instrumented: DOTNET-1800
                 new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/Oracle/{_fixture.TableName}/select", callCount = 2 },
-                new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/Oracle/{_fixture.TableName}/select", callCount = 2, metricScope = "WebTransaction/MVC/DefaultController/EnterpriseLibraryOracle"},
+                new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/Oracle/{_fixture.TableName}/select", callCount = 2, metricScope = "WebTransaction/MVC/DefaultController/OracleAsync"},
                 new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Oracle/insert", callCount = 1 },
                 new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/Oracle/{_fixture.TableName}/insert", callCount = 1 },
-                new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/Oracle/{_fixture.TableName}/insert", callCount = 1, metricScope = "WebTransaction/MVC/DefaultController/EnterpriseLibraryOracle"},
+                new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/Oracle/{_fixture.TableName}/insert", callCount = 1, metricScope = "WebTransaction/MVC/DefaultController/OracleAsync"},
                 new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Oracle/delete", callCount = 1 },
                 new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/Oracle/{_fixture.TableName}/delete", callCount = 1 },
-                new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/Oracle/{_fixture.TableName}/delete", callCount = 1, metricScope = "WebTransaction/MVC/DefaultController/EnterpriseLibraryOracle"},
-
-                // We are not checking callCount on Iterate metrics because they can be confusing in that calls like Open can result in calls to Read.
-                // This is particularly true for MySQL, but doing this for all vendors for consistency.
-                new Assertions.ExpectedMetric { metricName = @"DotNet/DatabaseResult/Iterate" },
-                new Assertions.ExpectedMetric { metricName = @"DotNet/DatabaseResult/Iterate", metricScope = "WebTransaction/MVC/DefaultController/EnterpriseLibraryOracle"}
+                new Assertions.ExpectedMetric { metricName = $@"Datastore/statement/Oracle/{_fixture.TableName}/delete", callCount = 1, metricScope = "WebTransaction/MVC/DefaultController/OracleAsync"}
+                // We don't currently have instrumentation for OracleDataReader.ReadAsync
+                //new Assertions.ExpectedMetric { metricName = @"DotNet/DatabaseResult/Iterate", callCount = 3 },
+                //new Assertions.ExpectedMetric { metricName = @"DotNet/DatabaseResult/Iterate", callCount = 3, metricScope = "WebTransaction/MVC/DefaultController/OracleAsync"}
             };
-
             var unexpectedMetrics = new List<Assertions.ExpectedMetric>
             {
                 // The datastore operation happened inside a web transaction so there should be no allOther metrics
-                new Assertions.ExpectedMetric { metricName = @"Datastore/allOther"},
-                new Assertions.ExpectedMetric { metricName = @"Datastore/Oracle/allOther"},
+                new Assertions.ExpectedMetric { metricName = @"Datastore/allOther" },
+                new Assertions.ExpectedMetric { metricName = @"Datastore/Oracle/allOther" },
 
                 // The operation metric should not be scoped because the statement metric is scoped instead
-                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Oracle/select", metricScope = "WebTransaction/MVC/DefaultController/EnterpriseLibraryOracle" },
-                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Oracle/insert", metricScope = "WebTransaction/MVC/DefaultController/EnterpriseLibraryOracle" },
-                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Oracle/delete", metricScope = "WebTransaction/MVC/DefaultController/EnterpriseLibraryOracle" }
+                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Oracle/select", metricScope = "WebTransaction/MVC/DefaultController/OracleAsync" },
+                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Oracle/insert", metricScope = "WebTransaction/MVC/DefaultController/OracleAsync" },
+                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Oracle/delete", metricScope = "WebTransaction/MVC/DefaultController/OracleAsync" }
             };
+
             var expectedTransactionTraceSegments = new List<string>
             {
                 "Datastore/statement/Oracle/user_tables/select"
@@ -94,18 +94,23 @@ namespace NewRelic.Agent.UnboundedIntegrationTests
                 "databaseDuration"
             };
 
+            var expectedTransactionTraceSegmentParameters = new List<Assertions.ExpectedSegmentParameter>
+            {
+                new Assertions.ExpectedSegmentParameter { segmentName = "Datastore/statement/Oracle/user_tables/select", parameterName = "sql", parameterValue = "SELECT DEGREE FROM user_tables WHERE ROWNUM <= ?"}
+            };
+
             var expectedSqlTraces = new List<Assertions.ExpectedSqlTrace>
             {
                 new Assertions.ExpectedSqlTrace
                 {
-                    TransactionName = "WebTransaction/MVC/DefaultController/EnterpriseLibraryOracle",
+                    TransactionName = "WebTransaction/MVC/DefaultController/OracleAsync",
                     Sql = "SELECT DEGREE FROM user_tables WHERE ROWNUM <= ?",
                     DatastoreMetricName = "Datastore/statement/Oracle/user_tables/select",
                     HasExplainPlan = false
                 },
                 new Assertions.ExpectedSqlTrace
                 {
-                    TransactionName = "WebTransaction/MVC/DefaultController/EnterpriseLibraryOracle",
+                    TransactionName = "WebTransaction/MVC/DefaultController/OracleAsync",
                     Sql = $"SELECT COUNT(*) FROM {_fixture.TableName}",
                     DatastoreMetricName = $"Datastore/statement/Oracle/{_fixture.TableName}/select",
 
@@ -113,7 +118,7 @@ namespace NewRelic.Agent.UnboundedIntegrationTests
                 },
                 new Assertions.ExpectedSqlTrace
                 {
-                    TransactionName = "WebTransaction/MVC/DefaultController/EnterpriseLibraryOracle",
+                    TransactionName = "WebTransaction/MVC/DefaultController/OracleAsync",
                     Sql = $"INSERT INTO {_fixture.TableName} (HOTEL_ID, BOOKING_DATE) VALUES (?, SYSDATE)",
                     DatastoreMetricName = $"Datastore/statement/Oracle/{_fixture.TableName}/insert",
 
@@ -121,7 +126,7 @@ namespace NewRelic.Agent.UnboundedIntegrationTests
                 },
                 new Assertions.ExpectedSqlTrace
                 {
-                    TransactionName = "WebTransaction/MVC/DefaultController/EnterpriseLibraryOracle",
+                    TransactionName = "WebTransaction/MVC/DefaultController/OracleAsync",
                     Sql = $"DELETE FROM {_fixture.TableName} WHERE HOTEL_ID = ?",
                     DatastoreMetricName = $"Datastore/statement/Oracle/{_fixture.TableName}/delete",
 
@@ -130,8 +135,8 @@ namespace NewRelic.Agent.UnboundedIntegrationTests
             };
 
             var metrics = _fixture.AgentLog.GetMetrics().ToList();
-            var transactionSample = _fixture.AgentLog.TryGetTransactionSample("WebTransaction/MVC/DefaultController/EnterpriseLibraryOracle");
-            var transactionEvent = _fixture.AgentLog.TryGetTransactionEvent("WebTransaction/MVC/DefaultController/EnterpriseLibraryOracle");
+            var transactionSample = _fixture.AgentLog.TryGetTransactionSample("WebTransaction/MVC/DefaultController/OracleAsync");
+            var transactionEvent = _fixture.AgentLog.TryGetTransactionEvent("WebTransaction/MVC/DefaultController/OracleAsync");
             var sqlTraces = _fixture.AgentLog.GetSqlTraces().ToList();
 
             NrAssert.Multiple(
@@ -145,7 +150,8 @@ namespace NewRelic.Agent.UnboundedIntegrationTests
                 () => Assertions.MetricsDoNotExist(unexpectedMetrics, metrics),
                 () => Assertions.TransactionTraceSegmentsExist(expectedTransactionTraceSegments, transactionSample),
                 () => Assertions.TransactionEventHasAttributes(expectedTransactionEventIntrinsicAttributes, TransactionEventAttributeType.Intrinsic, transactionEvent),
-                () => Assertions.SqlTraceExists(expectedSqlTraces, sqlTraces)
+                () => Assertions.SqlTraceExists(expectedSqlTraces, sqlTraces),
+                () => Assertions.TransactionTraceSegmentParametersExist(expectedTransactionTraceSegmentParameters, transactionSample)
             );
         }
     }
