+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/Oracle/OracleTests.cs
@@ -10,7 +10,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.Oracle
 {
     public class OracleTests : IClassFixture<RemoteServiceFixtures.OracleBasicMvcFixture>
     {
