+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/Postgres/PostgresAsyncLegacyAspPipelineTests.cs
@@ -10,7 +10,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.Postgres
 {
     public class PostgresAsyncLegacyAspPipelineTests : IClassFixture<RemoteServiceFixtures.BasicMvcApplication>
     {
