+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/Postgres/PostgresAsyncTests.cs
@@ -10,7 +10,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.Postgres
 {
     public class PostgresAsyncTests : IClassFixture<RemoteServiceFixtures.BasicMvcApplication>
     {
