+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/RabbitMq/RabbitMQTests.cs
@@ -8,7 +8,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.RabbitMq
 {
     public class RabbitMqTests : IClassFixture<RemoteServiceFixtures.RabbitMqBasicMvcFixture>
     {
